+++ packages/cli/CHANGELOG.md
@@ -7,7 +7,9 @@
 - Added VCS hooks management support.
   - Added `vcs.hooks` and `vcs.syncHooksOnRun` settings to `.moon/workspace.yml`.
   - Added `moon sync hooks` command.
-- Node.js
+- Updated `moon init` to:
+  - Detect an applicable VCS provider and set the `vcs.provider` setting.
+- **Node.js**
   - Moved syncing logic from `InstallNodeDeps` action to `SetupNodeTool` action. This includes
     syncing `packageManager`, `engines`, and version files. This should feel more natural.
 
@@ -59,8 +61,7 @@
 - Added a new action to the graph, `SyncWorkspace`, that'll be used for workspace-level checks.
 - Added `workspace.syncing` and `workspace.synced` webhooks.
 - Added `MOON_OUTPUT_STYLE` and `MOON_RETRY_COUNT` environment variables.
-
-- Rust
+- **Rust**
   - Improved Cargo workspace root and members detection.
 
 #### ⚙️ Internal